feat(wasi): provide default implementations for `WasiFile`
Additionally, as cleanup, remove duplicate implementations.
